pyproject.toml
[tool.poetry]
name = "openweather-fastapi"
version = "0.1.0"
description = ""
authors = ["levociraptor <leva.uspensky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.7.0"
anyio = "4.6.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
dnspython = "2.7.0"
email-validator = "2.2.0"
fastapi = "0.115.0"
fastapi-cli = "0.0.5"
h11 = "0.14.0"
httpcore = "1.0.6"
httptools = "0.6.1"
httpx = "0.27.2"
idna = "3.10"
iniconfig = "2.0.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
markdown-it-py = "3.0.0"
markupsafe = "3.0.1"
mdurl = "0.1.2"
mypy = "1.13.0"
mypy-extensions = "1.0.0"
orjson = "3.10.7"
packaging = "24.1"
pluggy = "1.5.0"
pydantic = "2.9.2"
pydantic-extra-types = "2.9.0"
pydantic-settings = "2.5.2"
pydantic-core = "2.23.4"
pygments = "2.18.0"
pytest = "8.3.3"
python-dotenv = "1.0.1"
python-multipart = "0.0.12"
pyyaml = "6.0.2"
rich = "13.9.2"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.38.6"
typer = "0.12.5"
typing-extensions = "4.12.2"
ujson = "5.10.0"
urllib3 = "2.2.3"
uvicorn = "0.31.1"
watchfiles = "0.24.0"
websockets = "13.1"
flake8 = "7.1.1"
flake8-bugbear = "24.8.19"
flake8-builtins = "2.5.0"
flake8-import-order = "0.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
